Update Vault main page, Add VaultType page
