+++ app/basicComponents/Tooltip.js
@@ -1,19 +1,25 @@
 // @flow
 import React, { PureComponent } from 'react';
 import styled from 'styled-components';
-import { tooltip } from '/assets/images';
+import { tooltip, tooltipWhite } from '/assets/images';
 import { smColors } from '/vars';
 
-const Wrapper = styled.div`
+const isDarkModeOn = localStorage.getItem('dmMode') === 'true';
+const icon = isDarkModeOn ? tooltipWhite : tooltip;
+
+const InnerWrapper = styled.div`
   display: none;
   position: absolute;
-  padding: 10px 15px;
+  top: ${({ top }) => top}px;
+  left: ${({ left }) => left}px;
+  width: ${({ width }) => width}px;
+  padding: 15px 20px;
   background-color: ${smColors.lightGray};
   border: 1px solid ${smColors.white};
   z-index: 10;
 `;
 
-const ToolTipIcon = styled.img`
+const InnerIcon = styled.img`
   position: absolute;
   top: 2px;
   left: 2px;
@@ -24,27 +30,39 @@ const ToolTipIcon = styled.img`
 const Text = styled.div`
   font-size: 10px;
   line-height: 13px;
-  text-transform: uppercase;
-  color: ${smColors.white};
+  color: ${isDarkModeOn ? smColors.white : smColors.black};
+`;
+
+const OuterIcon = styled.img`
+  width: 13px;
+  height: 13px;
+`;
+
+const Wrapper = styled.div`
+  position: relative;
+  margin-left: 5px;
+  &:hover ${InnerWrapper} {
+    display: block;
+  }
 `;
 
 type Props = {
-  className?: string,
-  text: string,
-  withIcon?: boolean
+  top: number,
+  left: number,
+  width: number,
+  text: string
 };
 
 class Tooltip extends PureComponent<Props> {
-  static defaultProps = {
-    withIcon: true
-  };
-
   render() {
-    const { className, withIcon, text } = this.props;
+    const { top, left, width, text } = this.props;
     return (
-      <Wrapper className={className}>
-        {withIcon && <ToolTipIcon src={tooltip} />}
-        <Text>{text}</Text>
+      <Wrapper>
+        <OuterIcon src={icon} />
+        <InnerWrapper top={top} left={left} width={width}>
+          <InnerIcon src={icon} />
+          <Text>{text}</Text>
+        </InnerWrapper>
       </Wrapper>
     );
   }
